Renaming folders to match GAS
https://eaflood.atlassian.net/browse/FUT-492
