+++ src/controllers/workflow.controller.test.js
@@ -5,10 +5,10 @@ import {
   workflowDetailController,
   workflowListController
 } from "./workflow.controller.js";
-import { workflowService } from "../service/workflow.service.js";
+import { workflowService } from "../services/workflow.service.js";
 import { workflowData1 } from "../../test/fixtures/workflow.js";
 
-vi.mock("../service/workflow.service.js");
+vi.mock("../services/workflow.service.js");
 
 describe("workflow.controller.js", () => {
   const mockRequest = {
